backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
openai==1.82.0
anthropic
python-dotenv==1.1.0
requests==2.32.3
reportlab==4.4.9
Pillow
supabase
